refactor: move SupabaseAuthenticationPrincipal to the user module
