+++ bandlog-user/src/main/java/net/effize/bandlog/auth/model/SupabaseAuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package net.effize.bandlog.shared.auth;
+package net.effize.bandlog.auth.model;
 
 
 public record SupabaseAuthenticationPrincipal(String supabaseUserId, String email) {
